fix: takeout separated index service of category module
