+++ providers/AppServiceInjectionProvider.cs
@@ -6,7 +6,6 @@ namespace onboarding_dotnet.Providers
     {
         public void InitServiceInjection()
         {
-            builder.Services.AddScoped<CategoryIndexService, CategoryIndexService>();
             builder.Services.AddScoped<CategoryService, CategoryService>();
             builder.Services.AddScoped<ProductIndexService, ProductIndexService>();
             builder.Services.AddScoped<ProductService, ProductService>();
